+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-
+### Added
+- Add cancelButton
 
 ## [2.0.2] - 2023-03-29
 ### Fixed